# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/get-notifications/{amount}/{page_no}": {
            "get": {
                "description": "Get first n notificatiobs n is the amount parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Get Notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "amount of notifications to be queried",
                        "name": "amount",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page to be gotten e.g if amount param is 8, 1 means first 8 products, 2 means from 9th to 16th notification",
                        "name": "page_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved notifications",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "something went wrong",
                        "schema": {
                            "$ref": "#/definitions/helpers.ServiceError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.ServiceError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}